replace index.js entry point
