+++ .dockerignore
@@ -4,6 +4,5 @@ npm-debug.log
 
 # Build files
 /index.js
-/icons.js
-/icons.mjs
-/icons.d.ts
+/index.mjs
+/index.d.ts
