+++ .gitignore
@@ -1,8 +1,7 @@
 # Files generated by build script
 /index.js
-/icons.js
-/icons.mjs
-/icons.d.ts
+/index.mjs
+/index.d.ts
 
 # Ignore all files in the icons folder
 icons/*
