+++ .npmignore
@@ -9,7 +9,6 @@
 !LICENSE.md
 !VERSIONING.md
 !index.js
+!index.mjs
 !index.d.ts
-!icons.js
-!icons.mjs
-!icons.d.ts
+!types.d.ts
