+++ package.json
@@ -59,7 +59,7 @@
   },
   "scripts": {
     "build": "node scripts/build/package.js",
-    "clean": "rimraf index.js icons.js icons.mjs icons.d.ts",
+    "clean": "rimraf index.js index.mjs index.d.ts",
     "format": "prettier --write .",
     "lint": "run-s our-lint jslint jsonlint svglint wslint",
     "our-lint": "node scripts/lint/ourlint.js",
